Schema class move to Schema.cs
